adaptive: make internal cache sizes size_t
The source size is a size_t and it represents a size in memory.

+++ modules/demux/adaptive/http/HTTPConnectionManager.h
@@ -102,8 +102,8 @@ namespace adaptive
                 AbstractConnection * reuseConnection(ConnectionParams &);
                 Downloader * getDownloadQueue(const AbstractChunkSource *) const;
                 std::list<HTTPChunkBufferedSource *> cache;
-                unsigned cache_total;
-                unsigned cache_max;
+                size_t cache_total;
+                size_t cache_max;
         };
     }
 }
